==> barbellrow_model/02. sklearn.ipynb <==
import mediapipe as mp
import cv2
import pandas as pd
import pickle

from sklearn.model_selection import train_test_split
from sklearn.linear_model import LogisticRegression, SGDClassifier
from sklearn.svm import SVC
from sklearn.neighbors import KNeighborsClassifier
from sklearn.tree import DecisionTreeClassifier
from sklearn.ensemble import RandomForestClassifier
from sklearn.naive_bayes import GaussianNB
from sklearn.metrics import precision_score, accuracy_score, f1_score, recall_score, confusion_matrix
from sklearn.preprocessing import StandardScaler
from sklearn.calibration import CalibratedClassifierCV

import warnings
warnings.filterwarnings('ignore')

# Drawing helpers
mp_drawing = mp.solutions.drawing_utils
mp_pose = mp.solutions.pose

def rescale_frame(frame, percent=50):
    '''
    Rescale a frame to a certain percentage compare to its original frame
    '''
    width = int(frame.shape[1] * percent/ 100)
    height = int(frame.shape[0] * percent/ 100)
    dim = (width, height)
    return cv2.resize(frame, dim, interpolation = cv2.INTER_AREA)


def describe_dataset(dataset_path: str):
    '''
    Describe dataset
    '''

    data = pd.read_csv(dataset_path)
    print(f"Headers: {list(data.columns.values)}")
    print(f'Number of rows: {data.shape[0]} \nNumber of columns: {data.shape[1]}\n')
    print(f"Labels: \n{data['label'].value_counts()}\n")
    print(f"Missing values: {data.isnull().values.any()}\n")
    
    duplicate = data[data.duplicated()]
    print(f"Duplicate Rows : {len(duplicate.sum(axis=1))}")

    return data


def round_up_metric_results(results) -> list:
    '''Round up metrics results such as precision score, recall score, ...'''
    return list(map(lambda el: round(el, 3), results))

# load dataset
df = describe_dataset("./lumbar_train.csv")

# Categorizing label
df.loc[df["label"] == "lumbar_correct", "label"] = 0
df.loc[df["label"] == "lumbar_error", "label"] = 1

# load dataset
df02 = describe_dataset("./torso_train.csv")

# Categorizing label
df02.loc[df02["label"] == "torso_correct", "label"] = 2
df02.loc[df02["label"] == "torso_error", "label"] = 3

# Extract features and class
X = df.drop("label", axis=1)
y = df["label"].astype("int")

# Extract features and class
a = df02.drop("label", axis=1)
b = df02["label"].astype("int")

sc = StandardScaler()
X = pd.DataFrame(sc.fit_transform(X))

sc02 = StandardScaler()
a = pd.DataFrame(sc02.fit_transform(a))

X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2, random_state=1234)
y_train

a_train, a_test, b_train, b_test = train_test_split(a, b, test_size=0.2, random_state=1234)
b_train

algorithms =[("LR", LogisticRegression()),
         ("SVC", SVC(probability=True)),
         ('KNN',KNeighborsClassifier()),
         ("DTC", DecisionTreeClassifier()),
         ("SGDC", CalibratedClassifierCV(SGDClassifier())),
         ("NB", GaussianNB()),
         ('RF', RandomForestClassifier()),]

models = {}
final_results = []

for name, model in algorithms:
    trained_model = model.fit(X_train, y_train)
    models[name] = trained_model

    # Evaluate model
    model_results = model.predict(X_test)

    p_score = precision_score(y_test, model_results, average=None, labels=[0, 1])
    a_score = accuracy_score(y_test, model_results)
    r_score = recall_score(y_test, model_results, average=None, labels=[0, 1])
    f1_score_result = f1_score(y_test, model_results, average=None, labels=[0, 1])
    cm = confusion_matrix(y_test, model_results, labels=[0, 1])
    final_results.append(( name,  round_up_metric_results(p_score), a_score, round_up_metric_results(r_score), round_up_metric_results(f1_score_result), cm))

# Sort results by F1 score
final_results.sort(key=lambda k: sum(k[4]), reverse=True)
pd.DataFrame(final_results, columns=["Model", "Precision Score", "Accuracy score", "Recall Score", "F1 score", "Confusion Matrix"])

algorithms02 =[("LR", LogisticRegression()),
         ("SVC", SVC(probability=True)),
         ('KNN',KNeighborsClassifier()),
         ("DTC", DecisionTreeClassifier()),
         ("SGDC", CalibratedClassifierCV(SGDClassifier())),
         ("NB", GaussianNB()),
         ('RF', RandomForestClassifier()),]

models02 = {}
final_results02 = []

for name, model in algorithms02:
    trained_model02 = model.fit(a_train, b_train)
    models02[name] = trained_model02

    # Evaluate model
    model_results02 = model.predict(a_test)

    p_score = precision_score(b_test, model_results02, average=None, labels=[2, 3])
    a_score = accuracy_score(b_test, model_results02)
    r_score = recall_score(b_test, model_results02, average=None, labels=[2, 3])
    f1_score_result = f1_score(b_test, model_results02, average=None, labels=[2, 3])
    cm = confusion_matrix(b_test, model_results02, labels=[2, 3])
    final_results02.append(( name,  round_up_metric_results(p_score), a_score, round_up_metric_results(r_score), round_up_metric_results(f1_score_result), cm))

# Sort results by F1 score
final_results02.sort(key=lambda k: sum(k[4]), reverse=True)
pd.DataFrame(final_results, columns=["Model", "Precision Score", "Accuracy score", "Recall Score", "F1 score", "Confusion Matrix"])

with open("./model/all_sklearn_lumbar.pkl", "wb") as f:
    pickle.dump(models, f)

with open("./model/all_sklearn_torso.pkl", "wb") as f:
    pickle.dump(models02, f)

with open("./model/RF_lumbar_model.pkl", "wb") as f:
    pickle.dump(models["RF"], f)

with open("./model/RF_torso_model.pkl", "wb") as f:
    pickle.dump(models02["RF"], f)

with open("./model/KNN_lumbar_model.pkl", "wb") as f:
    pickle.dump(models["KNN"], f)

with open("./model/KNN_torso_model.pkl", "wb") as f:
    pickle.dump(models02["KNN"], f)

with open("./model/SVC_lumbar_model.pkl", "wb") as f:
    pickle.dump(models["SVC"], f)

with open("./model/SVC_torso_model.pkl", "wb") as f:
    pickle.dump(models02["SVC"], f)

with open("./model/input_scaler_lumbar.pkl", "wb") as f:
    pickle.dump(sc, f)

with open("./model/input_scaler_torso.pkl", "wb") as f:
    pickle.dump(sc02, f)
